Fix lint: drop redundant HEAPU8 cast, now typed directly
The rebuilt pynterwasm.d.ts declares module.HEAPU8 as Uint8Array (a real
type, not the previous any) — the "as Uint8Array" assertion this line used
to need is now flagged by @typescript-eslint/no-unnecessary-type-assertion.

+++ src/engines/pvml/pynter/pynter-wasm.ts
@@ -53,7 +53,7 @@ export default async function init(props: Record<string, unknown> = {}): Promise
   alloc_heap(0x10000);
 
   const readReturnValue = (resPtr: number): PynterValue => {
-    const u8 = module.HEAPU8 as Uint8Array;
+    const u8 = module.HEAPU8;
     const dv = new DataView(u8.buffer);
     const type = dv.getUint32(resPtr, true);
     const raw32 = dv.getUint32(resPtr + 4, true);
